+++ Engines/NScumm.Scumm/IO/ResourceFile8.cs
@@ -299,14 +299,14 @@ namespace NScumm.Scumm.IO
             od.Name = name;
             var version = _reader.ReadUInt32();
             var numImnn = _reader.ReadInt32();
-            od.Position = new Point(_reader.ReadInt32(), _reader.ReadInt32());
+            od.Position = new Point((short) _reader.ReadInt32(), (short) _reader.ReadInt32());
             od.Width = (ushort)_reader.ReadInt32();
             od.Height = (ushort)_reader.ReadInt32();
             od.ActorDir = _reader.ReadInt32();
             od.Flags = (DrawBitmaps)_reader.ReadInt32();
             for (int i = 0; i < 15; i++)
             {
-                od.Hotspots.Add(new Point(_reader.ReadInt32(), _reader.ReadInt32()));
+                od.Hotspots.Add(new Point((short) _reader.ReadInt32(), (short) _reader.ReadInt32()));
             }
             return od;
         }
