+++ Engines/NScumm.Scumm/ScummEngine_Camera.cs
@@ -46,10 +46,10 @@ namespace NScumm.Scumm
             _camera.DestinationPosition.X = pos.X;
 
             if (VariableCameraMinX.HasValue && _camera.CurrentPosition.X < _variables[VariableCameraMinX.Value])
-                _camera.CurrentPosition.X = _variables[VariableCameraMinX.Value];
+                _camera.CurrentPosition.X = (short) _variables[VariableCameraMinX.Value];
 
             if (VariableCameraMaxX.HasValue && _camera.CurrentPosition.X > _variables[VariableCameraMaxX.Value])
-                _camera.CurrentPosition.X = _variables[VariableCameraMaxX.Value];
+                _camera.CurrentPosition.X = (short) _variables[VariableCameraMaxX.Value];
 
             if (VariableScrollScript.HasValue && _variables[VariableScrollScript.Value] != 0)
             {
@@ -95,12 +95,12 @@ namespace NScumm.Scumm
             Actor a = null;
             bool snapToX = /*_snapScroll ||*/ VariableCameraFastX.HasValue && _variables[VariableCameraFastX.Value] != 0;
 
-            _camera.CurrentPosition.X = (_camera.CurrentPosition.X & 0xFFF8);
+            _camera.CurrentPosition.X = (short) (_camera.CurrentPosition.X & 0xFFF8);
 
             if (VariableCameraMinX.HasValue && _camera.CurrentPosition.X < _variables[VariableCameraMinX.Value])
             {
                 if (snapToX)
-                    _camera.CurrentPosition.X = _variables[VariableCameraMinX.Value];
+                    _camera.CurrentPosition.X = (short) _variables[VariableCameraMinX.Value];
                 else
                     _camera.CurrentPosition.X += 8;
 
@@ -111,7 +111,7 @@ namespace NScumm.Scumm
             if (VariableCameraMaxX.HasValue && _camera.CurrentPosition.X > _variables[VariableCameraMaxX.Value])
             {
                 if (snapToX)
-                    _camera.CurrentPosition.X = _variables[VariableCameraMaxX.Value];
+                    _camera.CurrentPosition.X = (short) _variables[VariableCameraMaxX.Value];
                 else
                     _camera.CurrentPosition.X -= 8;
 
@@ -131,9 +131,9 @@ namespace NScumm.Scumm
                     if (snapToX)
                     {
                         if (t > 40 - 5)
-                            _camera.DestinationPosition.X = (actorx + 80);
+                            _camera.DestinationPosition.X = (short) (actorx + 80);
                         if (t < 5)
-                            _camera.DestinationPosition.X = (actorx - 80);
+                            _camera.DestinationPosition.X = (short) (actorx - 80);
                     }
                     else
                         _camera.MovingToActor = true;
@@ -147,10 +147,10 @@ namespace NScumm.Scumm
             }
 
             if (VariableCameraMinX.HasValue && _camera.DestinationPosition.X < _variables[VariableCameraMinX.Value])
-                _camera.DestinationPosition.X = _variables[VariableCameraMinX.Value];
+                _camera.DestinationPosition.X = (short) _variables[VariableCameraMinX.Value];
 
             if (VariableCameraMaxX.HasValue && _camera.DestinationPosition.X > _variables[VariableCameraMaxX.Value])
-                _camera.DestinationPosition.X = _variables[VariableCameraMaxX.Value];
+                _camera.DestinationPosition.X = (short) _variables[VariableCameraMaxX.Value];
 
             if (snapToX)
             {
@@ -185,11 +185,11 @@ namespace NScumm.Scumm
 
             if (_camera.CurrentPosition.X < (ScreenWidth / 2))
             {
-                _camera.CurrentPosition.X = (ScreenWidth / 2);
+                _camera.CurrentPosition.X = (short) (ScreenWidth / 2);
             }
             else if (_camera.CurrentPosition.X > (CurrentRoomData.Header.Width - (ScreenWidth / 2)))
             {
-                _camera.CurrentPosition.X = (CurrentRoomData.Header.Width - (ScreenWidth / 2));
+                _camera.CurrentPosition.X = (short) (CurrentRoomData.Header.Width - (ScreenWidth / 2));
             }
 
             _screenStartStrip = _camera.CurrentPosition.X / 8 - Gdi.NumStrips / 2;
